requirements.txt
# Core LangChain packages for document processing and RAG
langchain-community      # Document loaders (PyPDFLoader, Docx2txtLoader, TextLoader, WebBaseLoader)
langchain-core          # Core classes (Document, etc.)  
langchain               # Text splitter (RecursiveCharacterTextSplitter)

# Environment and configuration
python-dotenv           # Load environment variables from .env files

# Document processing libraries
pandas                  # Data manipulation for CSV/Excel files (.csv_loader.py, .xlsx_loader.py)
openpyxl               # Excel file support for pandas (.xlsx_loader.py)

# AI embeddings and vector database
google-generativeai    # Gemini API for embeddings (gemini_embedder.py)
pymilvus               # Milvus vector database client (milvus_vector_store.py)
